Implements Eran Henig's fix for Issues #110, #134, #142 by setting the android:configChanges="orientation" attribute in the property list for every POSIT activity.  In effect, this has the same effect on POSIT as would unchecking the "Auto-rotate-screen" setting in Android's Settings menu -- at least it seems to have the same effect on my Nexus One.
By adding the android:configChanges to the manifest for an Activity, you force Android to call the onConfigurationChanged() method.  By default this method is not called. Instead the Activity is restarted.  To handle a configuration change, the Activity should override the onConfigurationChanged method. By not overriding it in POSIT's activities, we are relying on the default behavior which seems to be merely to change the screen orientation.

Here's are links to the documentation on this: http://developer.android.com/reference/android/app/Activity.html#onConfigurationChanged%28android.content.res.Configuration%29 and http://developer.android.com/reference/android/R.attr.html#configChanges.